Port upstream's bloom threshold, pixel-centre picking (and fix the reversed pick ray), outline skipping of disabled entities, and Sky::setDepthWrite

+++ engine/src/platform/graphics/vertexBuffer.cpp
@@ -68,7 +68,7 @@ namespace visutwin::canvas
 
     bool VertexBuffer::setData(const std::vector<uint8_t>& data)
     {
-        if (data.size() != _numBytes) {
+        if (data.size() != static_cast<size_t>(_numBytes)) {
             spdlog::error("VertexBuffer: wrong initial data size: expected " +
                 std::to_string(_numBytes) + ", got " + std::to_string(data.size()));
             return false;
